Move image to top of contact page Fix missing data from form submission Refactor contact form feature
#18
