+++ .github/workflows/buildandtestprojects.yml
@@ -4,8 +4,6 @@
 name: .NET
 
 on:
-  push:
-    branches: [ "dev" ]
   pull_request:
     branches: [ "dev" ]
 
@@ -20,12 +18,6 @@ jobs:
       uses: actions/setup-dotnet@v4
       with:
         dotnet-version: 8.0.x
-    - uses: benday-inc/set-property-value-in-appsettings@main
-      with:
-        pathtosettingsfile: '${{ github.workspace }}/src/Web/appsettings.json'
-        keyname1: 'ContactFormEmailSettings'
-        keyname2: 'FromName'
-        valuetoset: '{{FROMNAME}}'
     - name: Restore dependencies
       run: dotnet restore
     - name: Build
