+++ src/Web/Components/Pages/Contact.razor
@@ -1,6 +1,6 @@
 ﻿@page "/Contact"
 @using System.ComponentModel.DataAnnotations
-@using Web.Features.Email
+@using Web.Features.ContactForm
 
 @inject IEmailService EmailService
 
@@ -48,10 +48,12 @@
                 <img src="/images/contact-kite.jpg" class="img-fluid"
                      alt="A kite in the sky with the string coming down" />
             </div>
-
-            @if (_displayForm)
-            {
-                <div class="col-lg-8">
+        </div>
+        <div class="row gy-4 mt-1">
+            <div class="col-lg-12">
+                <p>You can send me an email via this form. I promise I won't abuse it.</p>
+                @if (_displayForm)
+                {
                     <EditForm Model="_contactFormData"
                               OnValidSubmit="OnContactFormValidSubmit"
                               data-aos="fade-up"
@@ -96,20 +98,20 @@
                             <button type="submit">Send Message</button>
                         </div>
                     </EditForm>
-                </div>
-            }
-            else if (_isSending)
-            {
-                <div class="loading">
-                    Sending ...
-                </div>
-            }
-            else
-            {
-                <div class="sent-message">
-                    Your message has been sent. Thank you!
-                </div>
-            }
+                }
+                else if (_isSending)
+                {
+                    <div class="loading">
+                        Sending ...
+                    </div>
+                }
+                else
+                {
+                    <div class="sent-message">
+                        Your message has been sent. Thank you!
+                    </div>
+                }
+            </div>
         </div>
     </div>
 </section><!-- /Contact Section -->
@@ -126,7 +128,7 @@
         _displayForm = false;
 
         await EmailService.SendContactFormAsync(_contactFormData, _cancellationTokenSource.Token);
-        
+
         // I've added this artificial delay to show the "sending" message
         await Task.Delay(1_000);
