+++ src/Web/Features/ContactForm/ContactFormData.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Web.Features.Email;
+namespace Web.Features.ContactForm;
 
 public sealed class ContactFormData
 {
     [Required]
     public string Name { get; set; } = default!;
 
-    [Required, EmailAddress]
+    [Required(ErrorMessage = "The Email Address field is required."), EmailAddress]
     public string EmailAddress { get; set; } = default!;
 
     [Required]
@@ -15,4 +15,6 @@ public sealed class ContactFormData
 
     [Required]
     public string Message { get; set; } = default!;
+
+    public string EmailMessage => $"From:\r\n{Name}\r\n{EmailAddress}\r\n\r\n{Message}";
 }
